adding kernel to degridding

+++ src/functions.cu
@@ -1544,8 +1544,8 @@ __global__ void do_griddingGPU(float3 *uvw, cufftComplex *Vo, cufftComplex *Vo_g
    int k, j;
    if(i < visibilities)
    {
-    j = roundf(uvw[i].x / fabsf(deltau) + M/2);
-    k = roundf(uvw[i].y / fabsf(deltav) + N/2);
+    j = roundf(uvw[i].x / fabs(deltau) + M/2);
+    k = roundf(uvw[i].y / fabs(deltav) + N/2);
 
     if (k < M && j < N)
     {
